// models/ETechnology.ts
export enum ETechnology {
    REACT = 'React',
    REACTNATIVE = 'ReactNative',
    NODETS = 'NodeTS',
    NODEJS = 'NodeJS',
    EXPRESS = 'Express',
    JAVASCRIPT = 'JavaScript',
    TYPESCRIPT = 'TypeScript',
    JEST = 'Jest',
    MATERIALUI = 'MaterialUI',
    REDUX = 'Redux',
    FIREBASE = 'FireBase',
    FIREBASEAUTH = 'FireBaseAuth',
    NEXT = "Next",
    FORMIK = 'Formik',
    YUP = 'Yup',
    TESTINGLIBRARY = 'TestingLibrary',
    CLEANCODE = 'CleanCode',
    HEROKU = 'Heroku',
    VERCEL = 'Vercel',
    CYPRESSIO = 'CypressIO',
    ENZYME = 'Enzyme',
    GIMP = 'Gimp',
    POSTMAN = 'Postman',
    ADOBEXD = 'AdobeXD',
    FIGMA = 'Figma',
    MONGODB = 'MongoDB',
    Mongoos = 'Mongoose',
    TAILWINDCSS = 'TailwindCSS',
    JWT = 'JWT',
    JAVA = 'Java',
    ANGULAR = 'Angular',
    PHP = 'PHP',

}